refactor: pep8 change module name
