+++ address_separator/utils/modify/resplit_by_every_field_after_split.py
@@ -1,6 +1,6 @@
 import re
 
-from utils.modify.modifyExceptionAddressExpression import modify_exception_case_address2
+from utils.modify.modify_exception_address_expression import modify_exception_case_address2
 
 
 def re_split_by_every_fields(splittedAddressDataDictionaries: dict[str, list[str]]):
